add base file streaming

+++ src/storage.rs
@@ -6,6 +6,11 @@ pub type FileStream = ReaderStream<File>;
 /// File storage error
 #[derive(Debug, derive_more::From, derive_more::Display)]
 pub enum StorageError {
+    /// Attempting to write with read-only mode enabled
+    #[display("READ_ONLY_MODE_IS_ENABLED")]
+    #[from(skip)]
+    ReadOnlyEnabled,
+
     /// Incorrect key length
     #[display("BAD_FILE_KEY_LENGTH")]
     #[from(skip)]
@@ -18,6 +23,11 @@ pub enum StorageError {
     #[from(skip)]
     BadKeySpec,
 
+    /// The file was not found.
+    #[display("FILE_NOT_FOUND")]
+    #[from(skip)]
+    NotFound,
+
     // TODO: flatten io errors
     /// I/O error
     #[display("IO: {_0}")]
@@ -28,12 +38,17 @@ pub enum StorageError {
 #[derive(Debug, Clone)]
 pub struct Storage {
     root: std::path::PathBuf,
+
+    is_read_only: bool,
 }
 
 // TODO: Add key validation
 impl Storage {
     pub fn new(root: std::path::PathBuf) -> Self {
-        Storage { root }
+        Storage {
+            root,
+            is_read_only: false,
+        }
     }
 
     pub fn init(&self) -> Result<(), StorageError> {
@@ -43,6 +58,8 @@ impl Storage {
     }
 
     pub async fn get(&self, key: &str) -> Result<FileStream, StorageError> {
+        // todo: add base key validation
+
         let path = self.root.join(&key[..2]).join(&key[2..4]).join(key);
         let file = File::open(path).await?;
 
@@ -52,12 +69,20 @@ impl Storage {
     }
 
     pub async fn set(&self, key: &str) -> Result<(), StorageError> {
+        if self.is_read_only {
+            return Err(StorageError::ReadOnlyEnabled);
+        }
+
         let path = self.root.join(&key[..2]).join(&key[2..4]).join(key);
 
         todo!();
     }
 
     pub async fn remove(&self, key: &str) -> Result<(), StorageError> {
+        if self.is_read_only {
+            return Err(StorageError::ReadOnlyEnabled);
+        }
+
         let path = self.root.join(&key[..2]).join(&key[2..4]).join(key);
 
         tokio::fs::remove_file(path).await?;
